requirements.txt
cmake==3.27.0
dlib==19.24.2
face-recognition==1.2.3
numpy==1.25.2
opencv-python==4.8.0
Pillow==10.0.0

